+++ package.json
@@ -14,6 +14,7 @@
     "babylonjs": "^3.2.0-beta.3",
     "echarts": "^3.3.2",
     "element-ui": "^1.2.8",
+    "file-saver": "^1.3.8",
     "font-awesome": "^4.7.0",
     "normalize.css": "^5.0.0",
     "nprogress": "^0.2.0",
@@ -21,7 +22,8 @@
     "showdown": "1.7.1",
     "vue": "^2.2.2",
     "vue-router": "^2.3.0",
-    "vuex": "^2.0.0-rc.6"
+    "vuex": "^2.0.0-rc.6",
+    "xlsx": "^0.13.5"
   },
   "devDependencies": {
     "autoprefixer": "^6.7.2",
@@ -29,8 +31,8 @@
     "babel-core": "^6.22.1",
     "babel-loader": "^6.2.10",
     "babel-plugin-transform-runtime": "^6.22.0",
-    "babel-preset-env": "^1.2.1",
     "babel-polyfill": "^6.16.0",
+    "babel-preset-env": "^1.2.1",
     "babel-preset-es2015": "^6.0.0",
     "babel-preset-stage-2": "^6.22.0",
     "babel-register": "^6.22.0",
@@ -46,7 +48,6 @@
     "function-bind": "^1.0.2",
     "html-webpack-plugin": "^2.28.0",
     "http-proxy-middleware": "^0.17.3",
-    "webpack-bundle-analyzer": "^2.2.1",
     "json-loader": "^0.5.4",
     "mockjs": "^1.0.1-beta3",
     "node-sass": "^4.5.0",
@@ -55,6 +56,7 @@
     "ora": "^1.0.0",
     "rimraf": "^2.6.0",
     "sass-loader": "^6.0.0",
+    "script-loader": "^0.7.2",
     "semver": "^5.3.0",
     "shelljs": "^0.7.6",
     "url-loader": "^0.5.8",
@@ -62,6 +64,7 @@
     "vue-style-loader": "^2.0.0",
     "vue-template-compiler": "^2.2.4",
     "webpack": "^2.2.1",
+    "webpack-bundle-analyzer": "^2.2.1",
     "webpack-dev-middleware": "^1.10.0",
     "webpack-hot-middleware": "^2.16.1",
     "webpack-merge": "^2.6.1"
