+++ src/views/nav4/Page4.vue
@@ -1,106 +1,176 @@
 <template>
 	<section>
 		<!--工具条-->
-		<el-form ref="form" :model="form" @submit.prevent="onSubmit" style="margin:10px;">
-			<el-form :inline="true" :model="filters" class="demo-form-inline">
+		<el-collapse v-model="activeNames">
+			<el-collapse-item title="查询条件" name="1">
+                <el-form ref="form" :model="form" label-width="80px" @submit.prevent="onSubmit" style="margin:10px;width:60%;min-width:600px;">
+                    <el-row>
+                        <el-col :span="12">
+                            <el-form-item label="设备编号">
+                                <el-select v-model="form.deviceNo" placeholder="全部">
+                                    <!--<el-option label="全部" value="all"></el-option>-->
+                                    <!--<el-option label="蓄电池一" value="1"></el-option>-->
+                                    <!--<el-option label="蓄电池二" value="2"></el-option>-->
+                                    <el-option v-for="item in form.deviceNoItems" :key="item" v-bind:value="item.value">{{item.text}}</el-option>
+                                </el-select>
+                                <p>已选:{{form.deviceNo}}</p>
+                            </el-form-item>
+                        </el-col>
+                        <el-col :span="12">
+                            <el-form-item label-width="80px" label="开始时间">
+                                <el-date-picker v-model="form.startTime" type="datetime" format="yyyy-MM-dd HH:mm:ss" placeholder="选择日期时间">
+                                </el-date-picker>
+                            </el-form-item>
+                        </el-col>
+                    </el-row>
+                    <el-form-item label="重复设置">
+                        <el-checkbox-group v-model="form.repeat">
+                            <el-checkbox label="一次" name="repeat"></el-checkbox>
+                            <el-checkbox label="每天" name="repeat"></el-checkbox>
+                            <el-checkbox label="每周" name="repeat"></el-checkbox>
+                            <el-checkbox label="每季" name="repeat"></el-checkbox>
+                            <el-checkbox label="每年" name="repeat"></el-checkbox>
+                            <p>已选：{{form.repeat.join('|')}}</p>
+                        </el-checkbox-group>
+                    </el-form-item>
+                    <el-row>
+                        <el-col :span="12">
+                            <el-form-item label="类型选择">
+                                <el-checkbox-group v-model="form.type">
+                                    <el-checkbox label="充电" name="type"></el-checkbox>
+                                    <el-checkbox label="放电" name="type"></el-checkbox>
+                                    <el-checkbox label="内阻" name="type"></el-checkbox>
+                                </el-checkbox-group>
+                                <p>已选：{{form.type.join('|')}}</p>
+                            </el-form-item>
+                        </el-col>
+                        <el-col :span="12">
+                            <!--<el-form-item label="启用状态">-->
+                            <!--<el-checkbox-group v-model="form.state">-->
+                            <!--<el-checkbox label="已启用" name="type"></el-checkbox>-->
+                            <!--<el-checkbox label="未启用" name="type"></el-checkbox>-->
+                            <!--</el-checkbox-group>-->
+                            <!--<p>已选：{{form.state.join('|')}}</p>-->
+                            <!--</el-form-item>-->
+                            <el-form-item label="启用状态">
+                                <el-radio-group v-model="form.state">
+                                    <el-radio class="radio" :label="1">已启用</el-radio>
+                                    <el-radio class="radio" :label="0">未启用</el-radio>
+                                </el-radio-group>
+                            </el-form-item>
+                        </el-col>
+                    </el-row>
+                    <el-form :model="filters">
+                        <el-col :span="6">
+                            <el-form-item style="margin-left: 10px;">
+                                <el-input v-model="filters.siteName" placeholder="设备名称"></el-input>
+                            </el-form-item>
+                        </el-col>
+                        <el-col :span="4">
+                            <el-form-item style="margin-left: 30px;">
+                                <el-button type="primary" v-on:click="getUsers">查询</el-button>
+                            </el-form-item>
+                        </el-col>
+                    </el-form>
+                </el-form>
+			</el-collapse-item>
+		</el-collapse>
+
+        <div class="panel">
+            <div class="panel-title">
+                <span v-text="panelTitle"></span>
+                <div class="fr">
+                    <el-button @click.stop="on_refresh" size="small">
+                        <i class="fa fa-refresh"></i>
+                    </el-button>
+                    <el-button type="primary" @click="handleAdd" icon="plus" size="small">新增</el-button>
+                </div>
+            </div>
+            <div class="panel-body">
+                <!--列表-->
+                <el-table :data="users" highlight-current-row v-loading="listLoading" @selection-change="selsChange" style="width: 100%;">
+                    <el-table-column type="selection" width="55">
+                    </el-table-column>
+                    <el-table-column type="index" width="60">
+                    </el-table-column>
+                    <el-table-column prop="siteName" label="站点名称" width="120">
+                    </el-table-column>
+                    <el-table-column prop="deviceNo" label="设备号" width="100" sortable>
+                    </el-table-column>
+                    <el-table-column prop="startTime" label="开始时间" width="120" sortable>
+                    </el-table-column>
+                    <el-table-column prop="repeat" label="重复设置" min-width="120">
+                    </el-table-column>
+                    <el-table-column prop="state" label="启用状态" min-width="120" :formatter="formatState">
+                    </el-table-column>
+                    <el-table-column prop="type" label="类型" min-width="100">
+                    </el-table-column>
+                    <el-table-column label="操作" width="150">
+                        <template scope="scope">
+                            <el-button size="small" @click="handleEdit(scope.$index, scope.row)">编辑</el-button>
+                            <el-button type="danger" size="small" @click="handleDel(scope.$index, scope.row)">删除</el-button>
+                        </template>
+                    </el-table-column>
+                </el-table>
+
+                <!--工具条-->
+                <el-col :span="24" class="toolbar">
+                    <el-button type="danger" @click="batchRemove" :disabled="this.sels.length===0" class="fl">批量删除</el-button>
+                    <el-pagination @size-change="handleSizeChange" @current-change="handleCurrentChange" :current-page.sync="listQuery.curPage" :page-sizes="[10,20,30, 50]" :page-size="listQuery.pageSize" layout="total, sizes, prev, pager, next, jumper" :total="total" class="fr">
+                    </el-pagination>
+                </el-col>
+            </div>
+        </div>
+
+		<!--编辑界面-->
+		<el-dialog title="编辑" v-model="editFormVisible" :close-on-click-modal="false">
+			<el-form :model="editForm" label-width="80px" :rules="editFormRules" ref="editForm">
+				<el-form-item label="站点名称" prop="siteName">
+					<el-input v-model="editForm.siteName" auto-complete="off"></el-input>
+				</el-form-item>
+				<el-form-item label="设备号">
+					<el-input-number v-model="editForm.deviceNo" :min="0" :max="200"></el-input-number>
+				</el-form-item>
+				<el-form-item label="开始时间">
+					<el-date-picker type="date" placeholder="选择日期" v-model="editForm.startTime"></el-date-picker>
+				</el-form-item>
+                <!--<el-form-item label="重复设置">-->
+                    <!--<el-input type="textarea" v-model="editForm.repeat"></el-input>-->
+                <!--</el-form-item>-->
+				<el-form-item label="重复设置">
+					<el-checkbox-group v-model="editForm.repeat">
+						<el-checkbox label="一次" name="repeat"></el-checkbox>
+						<el-checkbox label="每天" name="repeat"></el-checkbox>
+						<el-checkbox label="每周" name="repeat"></el-checkbox>
+						<el-checkbox label="每季" name="repeat"></el-checkbox>
+						<el-checkbox label="每年" name="repeat"></el-checkbox>
+					</el-checkbox-group>
+				</el-form-item>
 				<el-row>
-					<el-col :xs="12" :sm="12" :md="12" :lg="12" style="margin-left: 12px;">
-						<el-form-item label="维护项">
-							<el-input v-model="filters.strTitle" placeholder="维护项" style="width: 160px;"></el-input>
+					<el-col :span="12">
+						<el-form-item label="类型选择">
+							<el-checkbox-group v-model="editForm.type">
+								<el-checkbox label="充电" name="type"></el-checkbox>
+								<el-checkbox label="放电" name="type"></el-checkbox>
+								<el-checkbox label="内阻" name="type"></el-checkbox>
+							</el-checkbox-group>
 						</el-form-item>
 					</el-col>
-					<el-col :xs="10" :sm="10" :md="10" :lg="10">
-						<el-form-item label="维护内容">
-							<el-input v-model="filters.strContent" placeholder="维护内容" style="width: 180px;"></el-input>
-						</el-form-item>
-					</el-col>
-					<el-col :xs="24" :sm="24" :md="24" :lg="12">
-						<el-form-item label-width="80px" label="创建时间" class="postInfo-container-item">
-							<el-date-picker
-									v-model="filters.createTime"
-									type="datetimerange"
-									:picker-options="pickerOptions2"
-									placeholder="选择时间范围"
-									align="right">
-							</el-date-picker>
-						</el-form-item>
-					</el-col>
-					<el-col :xs="24" :sm="24" :md="24" :lg="12">
-						<el-form-item label-width="80px" label="更新时间" class="postInfo-container-item">
-							<el-date-picker
-									v-model="filters.updateTime"
-									type="datetimerange"
-									:picker-options="pickerOptions2"
-									placeholder="选择时间范围"
-									align="right">
-							</el-date-picker>
-						</el-form-item>
-					</el-col>
-					<el-col :xs="24" :sm="24" :md="24" :lg="12" style="margin-left: 12px;">
-						<el-form-item>
-							<el-button type="primary" v-on:click="getMaintains" icon="search">查询</el-button>
+					<el-col :span="12">
+						<!--<el-form-item label="启用状态">-->
+							<!--<el-checkbox-group v-model="editForm.state">-->
+								<!--<el-checkbox label="已启用" name="type"></el-checkbox>-->
+								<!--<el-checkbox label="未启用" name="type"></el-checkbox>-->
+							<!--</el-checkbox-group>-->
+						<!--</el-form-item>-->
+						<el-form-item label="启用状态">
+							<el-radio-group v-model="editForm.state">
+								<el-radio class="radio" :label="1">已启用</el-radio>
+								<el-radio class="radio" :label="0">未启用</el-radio>
+							</el-radio-group>
 						</el-form-item>
 					</el-col>
 				</el-row>
-			</el-form>
-		</el-form>
-
-		<div class="panel">
-			<div class="panel-title">
-				<span v-text="panelTitle"></span>
-				<div class="fr">
-					<el-button @click.stop="on_refresh" size="small">
-						<i class="fa fa-refresh"></i>
-					</el-button>
-					<el-button type="primary" @click="handleAdd" icon="plus" size="small">新增</el-button>
-				</div>
-			</div>
-			<div class="panel-body">
-				<!--列表-->
-				<el-table :data="maintains" highlight-current-row v-loading="listLoading" @selection-change="selsChange">
-					<el-table-column type="selection" width="55">
-					</el-table-column>
-					<el-table-column type="index" width="60">
-					</el-table-column>
-					<el-table-column prop="strTitle" label="维护项">
-					</el-table-column>
-					<el-table-column prop="strContent" label="维护内容">
-					</el-table-column>
-					<el-table-column prop="createTime" label="创建时间">
-					</el-table-column>
-					<el-table-column prop="updateTime" label="更新时间">
-					</el-table-column>
-					<el-table-column label="操作" width="150">
-						<template scope="scope">
-							<el-button size="small" @click="handleEdit(scope.$index, scope.row)">编辑</el-button>
-							<el-button type="danger" size="small" @click="handleDel(scope.$index, scope.row)">删除</el-button>
-						</template>
-					</el-table-column>
-				</el-table>
-
-				<!--工具条-->
-				<el-col :span="24" class="toolbar">
-					<el-button type="danger" @click="batchRemove" :disabled="this.sels.length===0" class="fl">批量删除</el-button>
-					<el-pagination @size-change="handleSizeChange" @current-change="handleCurrentChange" :current-page.sync="listQuery.curPage" :page-sizes="[10,20,30, 50]" :page-size="listQuery.pageSize" layout="total, sizes, prev, pager, next, jumper" :total="total" class="fr">
-					</el-pagination>
-				</el-col>
-			</div>
-		</div>
-
-		<!--编辑界面-->
-		<el-dialog title="编辑维护项" v-model="editFormVisible" :close-on-click-modal="false">
-			<el-form :model="editForm" label-width="80px" :rules="editFormRules" ref="editForm">
-				<el-form-item label="名称" prop="strTitle" style="width: 450px;">
-					<el-input v-model="editForm.strTitle" auto-complete="off"></el-input>
-				</el-form-item>
-				<el-form-item label="维护内容" prop="strContent" style="width: 450px;">
-					<el-input
-							type="textarea"
-							:rows="2"
-							placeholder="请输入内容"
-							v-model="editForm.strContent">
-					</el-input>
-				</el-form-item>
 			</el-form>
 			<div slot="footer" class="dialog-footer">
 				<el-button @click.native="editFormVisible = false">取消</el-button>
@@ -109,23 +179,51 @@
 		</el-dialog>
 
 		<!--新增界面-->
-		<el-dialog title="新增维护项" v-model="addFormVisible" :close-on-click-modal="false" size="small">
+		<el-dialog title="新增" v-model="addFormVisible" :close-on-click-modal="false">
 			<el-form ref="addForm" :model="addForm" label-width="80px" :rules="addFormRules">
-				<!--<el-row :gutter="2">-->
-				<!--<el-col :xs="8" :sm="6" :md="5" :lg="5">-->
-				<el-form-item label="名称" prop="strTitle" style="width: 450px;">
-					<el-input v-model="addForm.strTitle" auto-complete="off"></el-input>
+				<el-form-item label="站点名称" prop="siteName">
+					<el-input v-model="addForm.siteName" auto-complete="off"></el-input>
 				</el-form-item>
-				<!--</el-col>-->
-				<!--</el-row>-->
-				<el-form-item label="维护内容" prop="strContent" style="width: 450px;">
-					<el-input
-							type="textarea"
-							:rows="2"
-							placeholder="请输入内容"
-							v-model="addForm.strContent">
-					</el-input>
+				<el-form-item label="设备号">
+					<el-input-number v-model="addForm.deviceNo" :min="0" :max="200"></el-input-number>
 				</el-form-item>
+				<el-form-item label="开始时间">
+					<el-date-picker type="date" placeholder="选择日期" v-model="addForm.startTime"></el-date-picker>
+				</el-form-item>
+				<el-form-item label="重复设置">
+					<el-checkbox-group v-model="addForm.repeat">
+						<el-checkbox label="一次" name="repeat"></el-checkbox>
+						<el-checkbox label="每天" name="repeat"></el-checkbox>
+						<el-checkbox label="每周" name="repeat"></el-checkbox>
+						<el-checkbox label="每季" name="repeat"></el-checkbox>
+						<el-checkbox label="每年" name="repeat"></el-checkbox>
+					</el-checkbox-group>
+				</el-form-item>
+				<el-row>
+					<el-col :span="12">
+						<el-form-item label="类型选择">
+							<el-checkbox-group v-model="addForm.type">
+								<el-checkbox label="充电" name="type"></el-checkbox>
+								<el-checkbox label="放电" name="type"></el-checkbox>
+								<el-checkbox label="内阻" name="type"></el-checkbox>
+							</el-checkbox-group>
+						</el-form-item>
+					</el-col>
+					<el-col :span="12">
+						<!--<el-form-item label="启用状态">-->
+							<!--<el-checkbox-group v-model="addForm.state">-->
+								<!--<el-checkbox label="已启用" name="type"></el-checkbox>-->
+								<!--<el-checkbox label="未启用" name="type"></el-checkbox>-->
+							<!--</el-checkbox-group>-->
+						<!--</el-form-item>-->
+						<el-form-item label="启用状态">
+							<el-radio-group v-model="editForm.state">
+								<el-radio class="radio" :label="1">已启用</el-radio>
+								<el-radio class="radio" :label="0">未启用</el-radio>
+							</el-radio-group>
+						</el-form-item>
+					</el-col>
+				</el-row>
 			</el-form>
 			<div slot="footer" class="dialog-footer">
 				<el-button @click.native="addFormVisible = false">取消</el-button>
@@ -138,124 +236,73 @@
 <script>
     import util from '../../common/js/util'
     //import NProgress from 'nprogress'
-    import { getMaintainListPage, removeMaintain, batchRemoveMaintain, editMaintain, addMaintain } from '../../api/api';
+    import { getUserListPage, removeUser, batchRemoveUser, editUser, addUser } from '../../api/api';
 
     export default {
         data() {
             return {
                 filters: {
-                    strTitle: '',
-                    strContent: '',
-                    createTime: '',
-                    updateTime: '',
-                    cStartTime: ''
+                    siteName: ''
                 },
-                panelTitle: '维护项列表',
-                pickerOptions1: {
-                    shortcuts: [{
-                        text: '今天',
-                        onClick(picker) {
-                            picker.$emit('pick', new Date());
-                        }
-                    }, {
-                        text: '昨天',
-                        onClick(picker) {
-                            const date = new Date();
-                            date.setTime(date.getTime() - 3600 * 1000 * 24);
-                            picker.$emit('pick', date);
-                        }
-                    }, {
-                        text: '一周前',
-                        onClick(picker) {
-                            const date = new Date();
-                            date.setTime(date.getTime() - 3600 * 1000 * 24 * 7);
-                            picker.$emit('pick', date);
-                        }
-                    }]
-                },
-                pickerOptions2: {
-                    shortcuts: [{
-                        text: '最近一周',
-                        onClick(picker) {
-                            const end = new Date();
-                            const start = new Date();
-                            start.setTime(start.getTime() - 3600 * 1000 * 24 * 7);
-                            picker.$emit('pick', [start, end]);
-                        }
-                    }, {
-                        text: '最近一个月',
-                        onClick(picker) {
-                            const end = new Date();
-                            const start = new Date();
-                            start.setTime(start.getTime() - 3600 * 1000 * 24 * 30);
-                            picker.$emit('pick', [start, end]);
-                        }
-                    }, {
-                        text: '最近三个月',
-                        onClick(picker) {
-                            const end = new Date();
-                            const start = new Date();
-                            start.setTime(start.getTime() - 3600 * 1000 * 24 * 90);
-                            picker.$emit('pick', [start, end]);
-                        }
-                    }]
-                },
-                maintains: [],
+                panelTitle: '数据列表',
+                users: [],
                 total: 0,
                 hello: true,
                 listLoading: false,
+                activeNames: ['1'],
                 sels: [],//列表选中列
 
                 form: {
-                    strTitle: '',
-                    strContent: '',
-                    cStartTime: '',
-                    cEndTime: '',
-                    uStartTime: '',
-                    uEndTime: ''
+                    siteName: '',
+                    deviceNo: '',
+                    deviceNoItems: [{text:'全部',value:'全部'},{text:'设备一',value:'设备一'},{text:'设备二',value:'设备二'}],
+                    startTime: '',
+					repeat: [],
+                    state: 0,
+                    type: ['放电','充电'],
+                    desc: ''
                 },
 
                 listQuery: {
                     curPage: 1,
                     limit: 20,
                     pageSize: 10,
-                    importance: undefined,
-                    title: undefined,
-                    type: undefined,
                     sort: '+id'
                 },
 
                 editFormVisible: false,//编辑界面是否显示
                 editLoading: false,
                 editFormRules: {
-                    strTitle: [
-                        { required: true, message: '请输入维护项名称', trigger: 'blur' }
-                    ],
-                    strContent: [
-                        { required: true, message: '请输入维护内容', trigger: 'blur' }
+                    name: [
+                        { required: true, message: '请输入姓名', trigger: 'blur' }
                     ]
                 },
                 //编辑界面数据
                 editForm: {
-                    strMaintainId: 0,
-                    strTitle: '',
-                    strContent: ''
+                    id: 0,
+                    siteName: '',
+                    deviceNo: '',
+                    startTime: '',
+                    repeat: [],
+                    state: 0,
+                    type: []
                 },
 
                 addFormVisible: false,//新增界面是否显示
                 addLoading: false,
                 addFormRules: {
-                    strTitle: [
-                        { required: true, message: '请输入维护项名称', trigger: 'blur' }
-                    ],
-                    strContent: [
-                        { required: true, message: '请输入维护内容', trigger: 'blur' }
+                    name: [
+                        { required: true, message: '请输入姓名', trigger: 'blur' }
                     ]
                 },
                 //新增界面数据
                 addForm: {
-                    strTitle: '',
-                    strContent: ''
+                    siteName: '',
+                    deviceNo: '',
+                    startTime: '',
+                    repeat: [],
+                    state: [],
+                    type: ['放电','充电']
                 }
 
             }
@@ -268,29 +315,39 @@
             //操作分页
             handleSizeChange(val) {
                 this.listQuery.pageSize = val;
-                this.getMaintains();
+                this.getUsers();
             },
             handleCurrentChange(val) {
                 this.listQuery.curPage = val;
-                this.getMaintains();
+                this.getUsers();
             },
             //刷新
             on_refresh(){
-                this.getMaintains();
+                this.getUsers();
             },
-            //获取维护项列表
-            getMaintains() {
+            //获取用户列表
+            getUsers() {
                 let para = {
                     curPage: this.listQuery.curPage,
                     pageSize: this.listQuery.pageSize,
-                    strTitle: this.filters.strTitle,
-                    strContent: this.filters.strContent
+                    siteName: this.filters.siteName
                 };
                 this.listLoading = true;
                 //NProgress.start();
-                getMaintainListPage(para).then((res) => {
+                getUserListPage(para).then((res) => {
                     this.total = res.data.total;
-                    this.maintains = res.data.maintains;
+                    this.users = res.data.users;
+                    for (var i = 0;  i < this.users.length; i++) {
+                        if(this.users[i].type.length > 1){
+                            var arr = [];
+                            for (var j = 0;  j < this.users[i].type.length-1; j++) {
+                                this.users[i].type[j] = this.users[i].type[j] + "|";
+								arr.push(this.users[i].type[j]);
+                            }
+                            arr.push(this.users[i].type[this.users[i].type.length-1]);
+                            this.users[i].type = arr;
+						}
+                    }
                     this.listLoading = false;
                     //NProgress.done();
                 });
@@ -302,15 +359,15 @@
                 }).then(() => {
                     this.listLoading = true;
                     //NProgress.start();
-                    let para = { strMaintainId: row.strMaintainId };
-                    removeMaintain(para).then((res) => {
+                    let para = { id: row.id };
+                    removeUser(para).then((res) => {
                         this.listLoading = false;
                         //NProgress.done();
                         this.$message({
                             message: '删除成功',
                             type: 'success'
                         });
-                        this.getMaintains();
+                        this.getUsers();
                     });
                 }).catch(() => {
 
@@ -320,13 +377,24 @@
             handleEdit: function (index, row) {
                 this.editFormVisible = true;
                 this.editForm = Object.assign({}, row);
+                let strs = []; //定义一数组
+				strs = this.editForm.type;
+                this.editForm.type = [];
+                for (let i = 0;  i < strs.length; i++) {
+                    const str = strs[i];
+                    this.editForm.type.push(str.substring(0,2));
+                }
             },
             //显示新增界面
             handleAdd: function () {
                 this.addFormVisible = true;
                 this.addForm = {
-                    strTitle: '',
-                    strContent: ''
+                    siteName: '',
+                    deviceNo: -1,
+                    startTime: '',
+                    repeat: ["每天"],
+                    state: [],
+                    type: ["放电", "内阻"]
                 };
             },
             //编辑
@@ -337,7 +405,8 @@
                             this.editLoading = true;
                             //NProgress.start();
                             let para = Object.assign({}, this.editForm);
-                            editMaintain(para).then((res) => {
+                            para.startTime = (!para.startTime || para.startTime == '') ? '' : util.formatDate.format(new Date(para.startTime), 'yyyy-MM-dd');
+                            editUser(para).then((res) => {
                                 this.editLoading = false;
                                 //NProgress.done();
                                 this.$message({
@@ -346,7 +415,7 @@
                                 });
                                 this.$refs['editForm'].resetFields();
                                 this.editFormVisible = false;
-                                this.getMaintains();
+                                this.getUsers();
                             });
                         });
                     }
@@ -360,7 +429,8 @@
                             this.addLoading = true;
                             //NProgress.start();
                             let para = Object.assign({}, this.addForm);
-                            addMaintain(para).then((res) => {
+                            para.startTime = (!para.startTime || para.startTime == '') ? '' : util.formatDate.format(new Date(para.startTime), 'yyyy-MM-dd');
+                            addUser(para).then((res) => {
                                 this.addLoading = false;
                                 //NProgress.done();
                                 this.$message({
@@ -369,7 +439,8 @@
                                 });
                                 this.$refs['addForm'].resetFields();
                                 this.addFormVisible = false;
-                                this.getMaintains();
+                                console.log("type " + para.type);
+                                this.getUsers();
                             });
                         });
                     }
@@ -380,21 +451,21 @@
             },
             //批量删除
             batchRemove: function () {
-                var ids = this.sels.map(item => item.strMaintainId).toString();
+                var ids = this.sels.map(item => item.id).toString();
                 this.$confirm('确认删除选中记录吗？', '提示', {
                     type: 'warning'
                 }).then(() => {
                     this.listLoading = true;
                     //NProgress.start();
                     let para = { ids: ids };
-                    batchRemoveMaintain(para).then((res) => {
+                    batchRemoveUser(para).then((res) => {
                         this.listLoading = false;
                         //NProgress.done();
                         this.$message({
                             message: '删除成功',
                             type: 'success'
                         });
-                        this.getMaintains();
+                        this.getUsers();
                     });
                 }).catch(() => {
 
@@ -402,7 +473,7 @@
             }
         },
         mounted() {
-            this.getMaintains();
+            this.getUsers();
         }
     }
